Refactoring of mapDispatchToProps
